requirements.txt
transformers>=4.35.0
torch>=2.4.0
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0
nltk>=3.8.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0
uvicorn>=0.23.2
pydantic-settings>=2.0.0
